# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Prompt Monitor Dashboard API
  description: API for monitoring AI prompts across multiple devices, providing real-time insights and compliance tracking.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://prompt-monitor-dashboard.vercel.app
    description: Production server
...
paths:
...
  /api/devices:
    get:
      tags:
        - Devices
      summary: List all devices
      description: Returns a list of all devices with computed statistics. Supports filtering and searching.
      operationId: listDevices
      parameters:
        - name: search
          in: query
          description: Search by hostname, ID, or IP address
          required: false
          schema:
            type: string
        - name: status
          in: query
          description: Filter by device status
          required: false
          schema:
            $ref: "#/components/schemas/DeviceStatus"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/devices/{id}:
    get:
      tags:
        - Devices
      summary: Get device details
      description: Returns detailed information about a specific device.
      operationId: getDevice
      parameters:
        - name: id
          in: path
          description: Device ID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/devices/register:
    post:
      tags:
        - Devices
      summary: Register a new device
      description: Register a new device with its public key and metadata. This is called by the native agent on first run.
      operationId: registerDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceRequest"
      responses:
        "200":
          description: Device registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterDeviceResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/devices/heartbeat:
    post:
      tags:
        - Devices
      summary: Send device heartbeat
      description: Update device last_heartbeat timestamp and optionally update metadata. Requires authentication.
      operationId: deviceHeartbeat
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeartbeatRequest"
      responses:
        "200":
          description: Heartbeat recorded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeartbeatResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/browsers/heartbeat:
    post:
      tags:
        - Browsers
      summary: Send browser heartbeat
      description: Update browser instance last_seen timestamp. Requires authentication.
      operationId: browserHeartbeat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrowserHeartbeatRequest"
      responses:
        "200":
          description: Heartbeat recorded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrowserHeartbeatResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/compliance/event:
    post:
      tags:
        - Compliance
      summary: Report compliance event
      description: Report a compliance violation or tampering attempt. Requires authentication.
      operationId: reportComplianceEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceEventRequest"
      responses:
        "200":
          description: Event reported successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceEventResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    DeviceStatus:
      type: string
      enum: [active, inactive, warning, error, compliance]
      description: Device status
...
    Device:
      type: object
      properties:
        id:
          type: string
          description: Device identifier
        hostname:
          type: string
          description: Device hostname
        status:
          $ref: "#/components/schemas/DeviceStatus"
        os:
          type: string
          description: Operating system
        ip_address:
          type: string
          description: Primary IP address
        last_seen:
          type: string
          format: date-time
          description: Last heartbeat timestamp (ISO 8601)
        browser_count:
          type: integer
          description: Number of browsers detected
        prompts_today:
          type: integer
          description: Number of prompts captured today
        total_prompts:
          type: integer
          description: Total number of prompts captured
        registered_at:
          type: string
          format: date-time
          nullable: true
          description: Device registration timestamp (ISO 8601)
      required:
        - id
        - hostname
        - status
        - os
        - ip_address
        - last_seen
        - browser_count
        - prompts_today
        - total_prompts
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
    RegisterDeviceRequest:
      type: object
      description: Request body for device registration
      properties:
        device_id:
          type: string
          description: Unique device identifier
        public_key_pem:
          type: string
          description: Device public key in PEM format for JWT verification
        hostname:
          type: string
          description: Device hostname
        macs:
          type: array
          items:
            type: string
          description: List of MAC addresses
        ips:
          type: array
          items:
            type: string
          description: List of IP addresses
        os:
          type: string
          description: Operating system
        browsers:
          type: array
          items:
            type: string
          description: List of detected browsers
      required:
        - device_id
        - public_key_pem
        - hostname
    RegisterDeviceResponse:
      type: object
      properties:
        status:
          type: string
          description: Registration status
        device_id:
          type: string
          description: Registered device ID
        message:
          type: string
          description: Success message
      required:
        - status
        - device_id
        - message
    HeartbeatRequest:
      type: object
      description: Request body for device heartbeat (optional metadata)
      properties:
        device_id:
          type: string
          description: Device ID (required until auth is implemented)
        device_metadata:
          type: object
          properties:
            hostname:
              type: string
            ips:
              type: array
              items:
                type: string
            macs:
              type: array
              items:
                type: string
            os:
              type: string
      required:
        - device_id
    HeartbeatResponse:
      type: object
      properties:
        status:
          type: string
          description: Heartbeat status
        device_id:
          type: string
          description: Device ID
        message:
          type: string
          description: Success message
      required:
        - status
        - device_id
        - message
...
    BrowserHeartbeatRequest:
      type: object
      description: Request body for browser heartbeat
      properties:
        instance_id:
          type: string
          description: Browser instance ID
      required:
        - instance_id
    BrowserHeartbeatResponse:
      type: object
      properties:
        status:
          type: string
          description: Heartbeat status
        instance_id:
          type: string
          description: Browser instance ID
        message:
          type: string
          description: Success message
      required:
        - status
        - instance_id
        - message
...
    ComplianceEventRequest:
      type: object
      description: Request body for compliance event report
      properties:
        instance_id:
          type: string
          nullable: true
          description: Browser instance ID
        event_type:
          type: string
          description: Type of compliance event
        severity:
          type: string
          enum: [info, warning, critical]
          description: Event severity
        detected_at:
          type: string
          format: date-time
          description: When the event was detected (ISO 8601)
        user_context:
          type: string
          nullable: true
          description: User context information
        details:
          type: object
          nullable: true
          description: Additional event details (JSONB)
          properties:
            site:
              type: string
            reason:
              type: string
      required:
        - event_type
        - severity
        - detected_at
    ComplianceEventResponse:
      type: object
      properties:
        status:
          type: string
          description: Report status
        event_id:
          type: string
          description: Created event ID
        message:
          type: string
          description: Success message
      required:
        - status
        - event_id
        - message