# src/generated/api/models/BrowserHeartbeatRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Request body for browser heartbeat
 */
export type BrowserHeartbeatRequest = {
  /**
   * Browser instance ID
   */
  instance_id: string;
};