# src/generated/api/models/BrowserHeartbeatResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BrowserHeartbeatResponse = {
  /**
   * Heartbeat status
   */
  status: string;
  /**
   * Browser instance ID
   */
  instance_id: string;
  /**
   * Success message
   */
  message: string;
};