# src/generated/api/models/HeartbeatResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type HeartbeatResponse = {
  /**
   * Heartbeat status
   */
  status: string;
  /**
   * Device ID
   */
  device_id: string;
  /**
   * Success message
   */
  message: string;
};